requirements.txt
python-dotenv==1.0.0
tavily-python==0.2.8
openai==1.3.3
python-telegram-bot==20.6
